Move down code to place of usage
